[C] fix warnings.

+++ aeron-driver/src/main/c/aeron_publication_image.c
@@ -461,7 +461,7 @@ int aeron_publication_image_send_pending_loss(aeron_publication_image_t *image)
         {
             const int32_t term_id = image->loss_term_id;
             const int32_t term_offset = image->loss_term_offset;
-            const int32_t length = image->loss_length;
+            const int32_t length = (int32_t)image->loss_length;
 
             aeron_acquire();
